+++ src/actions/index.js
@@ -1,0 +1,5 @@
+// @flow
+module.exports = {
+    query: require('./query'),
+    validate: require('./validate'),
+};
